fix() : resolve delete and update server side error in general department/ Department/ Branch/ Country/ City/ Town

+++ ServerLibrary/Repositores/Implementations/CityRepository.cs
@@ -19,13 +19,16 @@ public class CityRepository : IGenericRepositoryInterface<City>
 
     // CRUD Operations
 
-    public async Task<List<City>> GetAll() => await _context.Cities.ToListAsync();
+    public async Task<List<City>> GetAll() => await _context.Cities
+        .AsNoTracking()
+        .Include(c => c.Country)
+        .ToListAsync();
 
     public async Task<City?> GetById(int id) => await FindByIdAsync(id);
 
     public async Task<GeneralResponse> Create(City model)
     {
-        if (!await CheckName(model.Name)) return new GeneralResponse(false, nameof(City) + ConstantsResponse.Exit);
+        if (await CheckName(model.Name)) return Exited();
         await _context.Cities.AddAsync(model);
         await Commit();
         return Success();
@@ -35,15 +38,18 @@ public class CityRepository : IGenericRepositoryInterface<City>
     {
         var item = await FindByIdAsync(model.Id);
         if (item == null) return NotFound();
+        if (item.Name != model.Name && await CheckName(model.Name)) return Exited();
         item.Name = model.Name;
+        item.CountryId = model.CountryId;
         await Commit();
         return Success();
     }
 
     public async Task<GeneralResponse> DeleteById(int id)
     {
-        var item = await FindByIdAsync(id);
+        var item = await FindByIdAsyncWithChild(id);
         if (item == null) return NotFound();
+        if (item.Towns is not null && item.Towns.Any()) return HasChild();
         _context.Cities.Remove(item);
         await Commit();
         return Success();
@@ -53,11 +59,14 @@ public class CityRepository : IGenericRepositoryInterface<City>
 
     private GeneralResponse Success() => new GeneralResponse(true,ConstantsResponse.Success);
     private GeneralResponse NotFound() => new GeneralResponse(false, nameof(City) + ConstantsResponse.NotFound);
+    private GeneralResponse HasChild() => new GeneralResponse(false, nameof(City) + ConstantsResponse.HasChild + "of Branch");
+    private GeneralResponse Exited() => new GeneralResponse(false, nameof(City) + ConstantsResponse.Exit);
     private async Task Commit() => await _context.SaveChangesAsync();
     private async Task<City?> FindByIdAsync(int id) => await _context.Cities.FindAsync(id);
+    private async Task<City?> FindByIdAsyncWithChild(int id) => await _context.Cities.Include(c => c.Towns).FirstOrDefaultAsync(c => c.Id == id);
     private async Task<bool> CheckName(string name) 
     {
         var item = await _context.Cities.FirstOrDefaultAsync(_ => _.Name.ToLower() == name.ToLower());
-        return item == null ? false : true;
+        return item is null ? false : true;
     }
 }
