+++ ServerLibrary/Repositores/Implementations/GeneralDepartmentRepository.cs
@@ -27,7 +27,7 @@ namespace ServerLibrary.Repositores.Implementations
 
         public async Task<GeneralResponse> Create(GeneralDepartment model)
         {
-            if(await CheckName(model.Name)) return new GeneralResponse(false, nameof(GeneralDepartment) + ConstantsResponse.Exit);
+            if (await CheckName(model.Name)) return Exited();
             _context.GeneralDepartments.Add(model);
             await Commit();
             return Success();
@@ -36,7 +36,7 @@ namespace ServerLibrary.Repositores.Implementations
         {
             var item = await FindById(model.Id);
             if (item is null) return NotFound();
-            if (await CheckName(model.Name)) return new GeneralResponse(false, nameof(GeneralDepartment) + ConstantsResponse.Exit);
+            if (item.Name != model.Name && await CheckName(model.Name)) return Exited();
             item.Name = model.Name;
             await Commit();
             return Success();
@@ -44,8 +44,9 @@ namespace ServerLibrary.Repositores.Implementations
 
         public async Task<GeneralResponse> DeleteById(int id)
         {
-            var item = await FindById(id);
+            var item = await FindByIdWithChild(id);
             if (item is null) return NotFound();
+            if (item.Departments is not null && item.Departments.Any()) return HasChild();
             _context.GeneralDepartments.Remove(item);
             await Commit();
             return Success();
@@ -56,13 +57,16 @@ namespace ServerLibrary.Repositores.Implementations
         // reuseble propety and methode
         private async Task Commit() =>await _context.SaveChangesAsync();
         private async Task<GeneralDepartment?> FindById(int id) => await _context.GeneralDepartments.FindAsync(id);
+        private async Task<GeneralDepartment?> FindByIdWithChild(int id) => await _context.GeneralDepartments.Include(gd => gd.Departments).FirstOrDefaultAsync(gd => gd.Id == id);
         private static GeneralResponse NotFound() => new GeneralResponse(false, nameof(GeneralDepartment)+ConstantsResponse.NotFound);
+        private static GeneralResponse Exited() => new GeneralResponse(false, nameof(GeneralDepartment) + ConstantsResponse.Exit);
+        private static GeneralResponse HasChild() => new GeneralResponse(false, nameof(GeneralDepartment) + ConstantsResponse.HasChild + "of Branch");
         private static GeneralResponse Success()=> new GeneralResponse(true , ConstantsResponse.Success);
 
         private async Task<bool> CheckName(string name)
         {
             var item = await _context.GeneralDepartments.FirstOrDefaultAsync(_ => _.Name.ToLower() == name.ToLower());
-            return item == null ? false:  true;
+            return item is null ? false:  true;
         }
 
     }
